Integrated database.

+++ app/src/main/java/com/bluealeaf/dota2ticker/events/PassIdEvent.java
@@ -4,8 +4,8 @@ package com.bluealeaf.dota2ticker.events;
  * Created by samidh on 5/1/15.
  */
 public class PassIdEvent {
-    private int id;
-    public PassIdEvent(int id){
+    private long id;
+    public PassIdEvent(long id){
         this.id = id;
     }
     public PassIdEvent(){
@@ -13,7 +13,7 @@ public class PassIdEvent {
     }
 
 
-    public int getId(){
+    public long getId(){
         return id;
     }
